requirements.txt
fastapi
uvicorn[standard]
pydantic
groq
openai
python-dotenv
httpx